image repo fix 2

+++ mountain-climbing/src/main/java/com/example/mountainclimbing/service/ImageService.java
@@ -2,6 +2,9 @@ package com.example.mountainclimbing.service;
 import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.RequestParam;
 import com.example.mountainclimbing.model.Image;
@@ -17,9 +20,9 @@ public class ImageService {
 		this.imageRepository = imageRepository;
 	}
 
-	public Image createImage(Image image, HttpServletRequest req, HttpServletResponse res) {
+	public ResponseEntity<Image> createImage(Image image, HttpServletRequest req, HttpServletResponse res) {
 		imageRepository.save(image);
-		return image;
+		return new ResponseEntity<Image>(image,HttpStatus.OK);
 	}
 	
 	public Optional<Image> readImage(int id, HttpServletRequest req, HttpServletResponse res) {
@@ -27,17 +30,22 @@ public class ImageService {
 	}
 	
 	
-	public Image updateImage(int id,Image image, HttpServletRequest req, HttpServletResponse res) throws Exception {
+	public ResponseEntity<Image> updateImage(int id,Image image, HttpServletRequest req, HttpServletResponse res) {
 		if(imageRepository.existsById(id)) {
 			imageRepository.save(image);
-			return image;
+			return new ResponseEntity<Image>(image,HttpStatus.OK);
 		}
-		throw new Exception("Not found.");
+		return new ResponseEntity<Image>(image,HttpStatus.NOT_FOUND);
+		
 	}
 	
 	
-	public void deleteImage(@RequestParam int id, HttpServletRequest req, HttpServletResponse res) {
+	public ResponseEntity<Void> deleteImage(@RequestParam int id, HttpServletRequest req, HttpServletResponse res) {
+		if(imageRepository.existsById(id)) {
 		imageRepository.deleteById(id);
+		return new ResponseEntity<Void>(HttpStatus.OK);
+		}
+		return new ResponseEntity<Void>(HttpStatus.NOT_FOUND);
 	}
 
 }
